chore: split FFI related code to another file.
For modularity.
The code also only used in windows
